// src/api/models/userProfileUpdateRequestGender.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FlowNews API
 * OpenAPI spec version: v1
 */

export type UserProfileUpdateRequestGender = typeof UserProfileUpdateRequestGender[keyof typeof UserProfileUpdateRequestGender];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const UserProfileUpdateRequestGender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE',
  NON_BINARY: 'NON_BINARY',
  PREFER_NOT_TO_SAY: 'PREFER_NOT_TO_SAY',
} as const;